+++ DigitalLibrary.API/Controllers/LibraryController.cs
@@ -3,6 +3,7 @@ using DigitalLibrary.API.DTOs;
 using DigitalLibrary.API.Services;
 using Microsoft.AspNetCore.Mvc;
 using DigitalLibrary.API.Common;
+using Microsoft.AspNetCore.Authorization;
 
 namespace DigitalLibrary.API.Controllers;
 
@@ -15,6 +16,7 @@ namespace DigitalLibrary.API.Controllers;
 /// </remarks>
 [ApiController]
 [Route("api/users/{userId:int}/library")]
+[Authorize]
 public class LibraryController : ControllerBase
 {
     private readonly ILibraryService _libraryService;
@@ -27,15 +29,16 @@ public class LibraryController : ControllerBase
     /// <summary>
     /// Retrieves the library associated with a specific user.
     /// </summary>
-    /// <param name="userId">The ID of the user whose library is being retrieved.</param>
     /// <returns>
     /// A <see cref="LibraryReadDTO"/> object representing the user's library.
     /// </returns>
     /// <response code="200">Successfully retrieved the user's library.</response>
     /// <response code="404">No library found for the specified user.</response>
     [HttpGet]
-    public async Task<ActionResult<LibraryReadDTO>> GetLibrary([FromRoute] int userId)
+    public async Task<ActionResult<LibraryReadDTO>> GetLibrary()
     {
+        if (!UserClaimsHelper.TryGetUserId(this, out int userId)) return Unauthorized();
+
         var result = await _libraryService.GetLibraryByUserIdAsync(userId);
         if(!result.IsSuccess)
         {
@@ -52,7 +55,6 @@ public class LibraryController : ControllerBase
     /// <summary>
     /// Creates a new library for a specific user.
     /// </summary>
-    /// <param name="userId">The ID of the user for whom the library will be created.</param>
     /// <param name="libraryCreateDTO">The data used to create the new library.</param>
     /// <returns>
     /// The newly created <see cref="LibraryReadDTO"/> object.
@@ -61,8 +63,10 @@ public class LibraryController : ControllerBase
     /// <response code="404">The specified user was not found.</response>
     /// <response code="400">The provided library data is invalid.</response>
     [HttpPost]
-    public async Task<ActionResult<LibraryReadDTO>> CreateLibrary([FromRoute] int userId, [FromBody] LibraryCreateDTO libraryCreateDTO)
+    public async Task<ActionResult<LibraryReadDTO>> CreateLibrary([FromBody] LibraryCreateDTO libraryCreateDTO)
     {
+        if (!UserClaimsHelper.TryGetUserId(this, out int userId)) return Unauthorized();
+
         var result = await _libraryService.AddLibraryAsync(libraryCreateDTO, userId);
         if (!result.IsSuccess)
         {
@@ -80,12 +84,13 @@ public class LibraryController : ControllerBase
     /// <summary>
     /// Deletes the library associated with a specific user.
     /// </summary>
-    /// <param name="userId">The ID of the user whose library will be deleted.</param>
     /// <response code="204">Successfully deleted the user's library.</response>
     /// <response code="404">No library found for the specified user.</response>
     [HttpDelete]
-    public async Task<IActionResult> DeleteLibrary([FromRoute] int userId)
+    public async Task<IActionResult> DeleteLibrary()
     {
+        if (!UserClaimsHelper.TryGetUserId(this, out int userId)) return Unauthorized();
+
         var result = await _libraryService.DeleteLibraryAsync(userId);
         if (!result.IsSuccess)
         {
@@ -103,13 +108,14 @@ public class LibraryController : ControllerBase
     /// <summary>
     /// Updates the library associated with a specific user.
     /// </summary>
-    /// <param name="userId">The ID of the user whose library is being updated.</param>
     /// <param name="libraryUpdateDTO">The updated library information.</param>
     /// <response code="204">Successfully updated the user's library.</response>
     /// <response code="404">No library found for the specified user.</response>
     [HttpPut]
-    public async Task<IActionResult> UpdateLibrary([FromRoute] int userId, [FromBody] LibraryUpdateDTO libraryUpdateDTO)
+    public async Task<IActionResult> UpdateLibrary([FromBody] LibraryUpdateDTO libraryUpdateDTO)
     {
+        if (!UserClaimsHelper.TryGetUserId(this, out int userId)) return Unauthorized();
+
         var result = await _libraryService.UpdateLibraryAsync(libraryUpdateDTO, userId);
         if (!result.IsSuccess)
         {
